modules/libmrpc: drop deprecated funs from mrpc.h
